Handle all flag-prodcut-bands for GetFeatureInfo requests.
